requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
pymongo==3.11.4
geopy==2.2.0
python-dotenv==0.19.0
dnspython==2.2.1
